# curriculum-elkas-backend/config/swaggerDocument.yaml
swagger: "2.0"
info:
  description: "API do software: Curriculum-Elkas: Gerador de currículo online"
  version: 1.0.0
  title: Curriculum-Elkas API
  contact:
    email: gabriellima.silva96@gmail.com
  license:
    name: MIT
    url: "https://opensource.org/licenses/MIT"
host: api.curriculumelkas.com
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
basePath: /api
...
schemes:
  - http
paths:
  "/users/{username}":
    get:
      tags:
        - Users
      summary: Buscar o usuário por username
      description: Buscar informações sobre o usuário pelo username
      operationId: user.index
      consumes:
        - application/json
      produces:
        - application/json
      security:
        - Bearer: []
      parameters:
        - in: path
          name: username
          description: O username que precisa ser buscado
          required: true
          type: string
      responses:
        "200":
          description: Sucesso na operação!
          schema:
            $ref: "#/definitions/User"
        "403":
          description: Permissão negada para acessar esse usuário!
          schema:
            $ref: "#/definitions/ApiResponse"
        "404":
          description: Usuário não encontrado!
          schema:
            $ref: "#/definitions/ApiResponse"
    put:
      tags:
        - Users
      summary: Atualizar o usuário por username
      description: Atualizar informações sobre o usuário pelo username
      operationId: user.update
      security:
        - Bearer: []
      produces:
        - application/xml
        - application/json
      parameters:
        - name: username
          in: path
          description: O username que precisa ser buscado
          required: true
          type: string
        - in: body
          name: body
          description: Objeto user atualizado
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
              username:
                type: string
              email:
                type: string
              currentPassword:
                type: string
              newPassword:
                type: string
              personalInformation:
                type: object
                properties:
                  emails:
                    type: array
                    items:
                      type: string
                  phones:
                    type: array
                    items:
                      type: string
                  dataOfbirth:
                    type: string
                    format: date
                  maritalStatus:
                    type: string
                  academicDegree:
                    type: array
                    items:
                      type: object
                      properties:
                        institution:
                          type: string
                        degree:
                          type: string
                        course:
                          type: string
                        initialization:
                          type: string
                          format: date
                        conclusion:
                          type: string
                          format: date
                  address:
                    type: object
                    properties:
                      publicArea:
                        type: string
                      district:
                        type: string
                      city:
                        type: string
                      postalCode:
                        type: string
                      country:
                        type: string
                      state:
                        type: string
      responses:
        "200":
          description: Sucesso na operação!
          schema:
            $ref: "#/definitions/User"
        "400":
          description: Já existe um usuário com o e-mail cadastrado!
          schema:
            $ref: "#/definitions/ApiResponse"
        "401":
          description: Senha atual inválida
          schema:
            $ref: "#/definitions/ApiResponse"
        "403":
          description: Permissão negada para acessar esse usuário!
          schema:
            $ref: "#/definitions/ApiResponse"
        "404":
          description: Usuário não encontrado!
          schema:
            $ref: "#/definitions/ApiResponse"
        "409":
          description: Já existe um usuário com o username cadastrado!
          schema:
            $ref: "#/definitions/ApiResponse"
    delete:
      tags:
         - Users
      summary: Remove um usuário e os respectivos currículos
      description: Apaga a Conta no Sistema
      operationId: user.destroy
      consumes:
        - application/json
      produces:
        - application/json
      security:
        - Bearer: []
      parameters:
        - name: username
          in: path
          description: O username do usuário
          required: true
          type: string
      responses:
        '200':
          description: Sucesso!
        '404':
          description: Usuário não encontrado!
        '500':
          description: Erro interno no servidor!
...
  "/curriculums/:id":
    get:
      tags:
        - Curriculums
      summary: Buscar um curriculum por id
      description: Buscar um curriculum do usuário por id
      operationId: curriculum.show
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID do curriculum
          required: true
          type: string
      responses:
        "200":
          description: Sucesso!
          schema:
            $ref: "#/definitions/Curriculum"
        "404":
          description: Currículo não encontrado!
        "500":
          description: Error interno no servidor!
    put:
      tags:
        - Curriculums
      summary: Editar um curriculum por id
      description: Editar um curriculum do usuário por id
      operationId: curriculum.update
      consumes:
        - application/json
      produces:
        - application/json
      security:
        - Bearer: []
      parameters:
        - name: id
          in: path
          description: ID do curriculum
          required: true
          type: string
        - in: body
          name: body
          description: Objeto curriculum atualizado
          required: true
          schema:
            type: object
            properties:
              template:
                type: string
              name:
                type: string
              emails:
                type: array
                items:
                  type: string
              phones:
                type: array
                items:
                  type: string
              dataOfbirth:
                type: string
                format: date
              maritalStatus:
                type: string
              academicDegree:
                type: array
                items:
                  type: object
                  properties:
                    institution:
                      type: string
                    degree:
                      type: string
                    course:
                      type: string
                    initialization:
                      type: string
                      format: date
                    conclusion:
                      type: string
                      format: date
              address:
                type: object
                properties:
                  publicArea:
                    type: string
                  district:
                    type: string
                  city:
                    type: string
                  postalCode:
                    type: string
                  country:
                    type: string
                  state:
                    type: string
              goals:
                type: string
              experiences:
                type: array
                items:
                  type: object
                  properties:
                    office:
                      type: string
                    location:
                      type: string
                    company:
                      type: string
                    description:
                      type: string
                    type:
                      type: string
                    initialization:
                      type: string
                      format: date
                    conclusion:
                      type: string
                      format: date
              languages:
                type: array
                items:
                  type: object
                  properties:
                    language:
                      type: string
                    fluency:
                      type: string
              skills:
                type: array
                items:
                  type: object
                  properties:
                    title:
                      type: string
                    description:
                      type: string
                    other:
                      type: string
      responses:
        "200":
          description: Sucesso!
          schema:
            $ref: "#/definitions/Curriculum"
        "404":
          description: Currículo não encontrado!
        "500":
          description: Erro interno no servidor!
    delete:
      tags:
        - Curriculums
      summary: Remove um curriculum
      description: Remove um curriculum do usuário
      operationId: curriculum.destroy
      consumes:
        - application/json
      produces:
        - application/json
      security:
        - Bearer: []
      parameters:
        - name: id
          in: path
          description: ID do curriculum
          required: true
          type: string
      responses:
        "200":
          description: Sucesso!
        '404':
          description: Curriculum não encontrado!
        '500':
          description: Erro interno no servidor!
definitions:
...
  ApiResponse:
    type: object
    properties:
      status:
        type: number
      error:
        type: string
  User:
    type: object
    properties:
      name:
        type: string
      username:
        type: string
      email:
        type: string
      password:
        type: string
      personalInformation:
        type: object
        properties:
          emails:
            type: array
            items:
              type: string
          phones:
            type: array
            items:
              type: string
          dataOfbirth:
            type: string
            format: date
          maritalStatus:
            type: string
          academicDegree:
            type: array
            items:
              type: object
              properties:
                institution:
                  type: string
                degree:
                  type: string
                course:
                  type: string
                initialization:
                  type: string
                  format: date
                conclusion:
                  type: string
                  format: date
          address:
            type: object
            properties:
              publicArea:
                type: string
              district:
                type: string
              city:
                type: string
              postalCode:
                type: string
              country:
                type: string
              state:
                type: string
...
  Curriculum:
    type: object
    properties:
      template:
        type: string
      name:
        type: string
      emails:
        type: array
        items:
          type: string
      phones:
        type: array
        items:
          type: string
      dataOfbirth:
        type: string
        format: date
      maritalStatus:
        type: string
      academicDegree:
        type: array
        items:
          type: object
          properties:
            institution:
              type: string
            degree:
              type: string
            course:
              type: string
            initialization:
              type: string
              format: date
            conclusion:
              type: string
              format: date
      address:
        type: object
        properties:
          publicArea:
            type: string
          district:
            type: string
          city:
            type: string
          postalCode:
            type: string
          country:
            type: string
          state:
            type: string
      goals:
        type: string
      experiences:
        type: array
        items:
          type: object
          properties:
            office:
              type: string
            location:
              type: string
            company:
              type: string
            description:
              type: string
            type:
              type: string
            initialization:
              type: string
              format: date
            conclusion:
              type: string
              format: date
      languages:
        type: array
        items:
          type: object
          properties:
            language:
              type: string
            fluency:
              type: string
      skills:
        type: array
        items:
          type: object
          properties:
            title:
              type: string
            description:
              type: string
            other:
              type: string